refactor/ getting rid of middleware
